Fix trait
