+++ src/Exceptions/ShouldBeSkippedMetricExceptionInterface.php
@@ -2,9 +2,7 @@
 
 namespace AvtoDev\AppMetrics\Exceptions;
 
-use Throwable;
-
-interface ShouldBeSkippedMetricExceptionInterface extends Throwable
+interface ShouldBeSkippedMetricExceptionInterface extends \Throwable
 {
     //
 }
