+++ src/Traits/ThrowableToExceptionTrait.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace AvtoDev\AppMetrics\Traits;
+
+use Exception;
+use Throwable;
+
+trait ThrowableToExceptionTrait
+{
+    /**
+     * @param Throwable $e
+     *
+     * @return Exception
+     */
+    protected function convertThrowableToException(Throwable $e): Exception
+    {
+        if ($e instanceof Exception) {
+            return $e;
+        }
+
+        return new \RuntimeException($e->getMessage(), $e->getCode(), $e);
+    }
+}
